update: Schema v2. replace lowFps(boolean) to powerLevel(int).

+++ src/index.ts
@@ -1,9 +1,12 @@
 import fs from "fs";
+import path from "path";
 // module to parse command line arguments
 import yargs from "yargs";
 
 import { OptionValues, program } from "commander";
 import Realm from "realm";
+
+import { SCHEMA_VERSION } from "./models/RealmDataConst";
 
 import { RealmManufacturerCatalog } from "./schemas/RealmManufacturerCatalog";
 import { RealmGunCatalog } from "./schemas/RealmGunCatalog";
@@ -16,26 +19,38 @@ import { createGunCatalog } from "./createGunCatalog";
 const REALM_FILE_NAME = "catalog_data.realm";
 
 async function main(options: OptionValues, directory: string): Promise<boolean> {
-    const realm = await (async () => {;
-        const schema = [RealmManufacturerCatalog, RealmGunCatalog, RealmShootingRuleCatalog]
+    let realm: Realm;
 
-        if (options.test) {
-            return await Realm.open({
-                inMemory: true,
-                schema: schema,
-            });
-        } else {
-            if (fs.existsSync(options.output)) {
-                console.log(`ERROR: Output file ${options.output} already exists.`);
-                return false;
+    try {
+        realm = await (async () => {;
+            const schema = [RealmManufacturerCatalog, RealmGunCatalog, RealmShootingRuleCatalog]
+    
+            if (options.test) {
+                return await Realm.open({
+                    inMemory: true,
+                    schema: schema,
+                    schemaVersion: SCHEMA_VERSION,
+                });
+            } else {
+                if (fs.existsSync(options.output)) {
+                    throw new Error(`ERROR: Output file ${options.output} already exists.`);
+                }
+    
+                return await Realm.open({
+                    path: options.output,
+                    schema: schema,
+                    schemaVersion: SCHEMA_VERSION,
+                });
             }
-
-            return await Realm.open({
-                path: options.output,
-                schema: schema,
-            });
+        })();    
+    } catch (e) {
+        if (e instanceof Error) {
+            console.error(e.message);
+        } else {
+            console.error(`Unknown error: ${JSON.stringify(e)}`)
         }
-    })();
+        return false
+    }
 
     if (realm instanceof Realm) {
         try {
@@ -49,19 +64,19 @@ async function main(options: OptionValues, directory: string): Promise<boolean> 
             }
 
             file = `${directory}RealmManufacturerCatalog.csv`;
-            console.log(`\nCreating manufacturer catalog from ${file}`)
+            console.log(`\nWriting manufacturer catalog from ${file}`)
             success = createManufacturerCatalog(realm, `${file}`);
-            console.log(`${success ? "Succeeded" : "FAILED"}`);
+            console.log(`${success ? "Done" : "FAILED"}`);
 
             file = `${directory}RealmGunCatalog.csv`;
-            console.log(`\nCreating gun catalog from ${file}`)
+            console.log(`\nWriting gun catalog from ${file}`)
             success = createGunCatalog(realm, `${file}`);
-            console.log(`${success ? "Succeeded" : "FAILED"}`);
+            console.log(`${success ? "Done" : "FAILED"}`);
 
             file = `${directory}RealmShootingRuleCatalog.csv`;
-            console.log(`\nCreating shooting rule catalog from ${file}`)
+            console.log(`\nWriting shooting rule catalog from ${file}`)
             success = createShootingRuleCatalog(realm, `${file}`);
-            console.log(`${success ? "Succeeded" : "FAILED"}`);
+            console.log(`${success ? "Done" : "FAILED"}`);
 
             realm.close();
         } catch(e) {
@@ -71,7 +86,9 @@ async function main(options: OptionValues, directory: string): Promise<boolean> 
                 console.error(`Unknown error: ${JSON.stringify(e)}`)
             }
             realm.close();
-            fs.unlinkSync(options.output); // delete output file
+            if (!options.test) {
+                fs.unlinkSync(options.output); // delete output file
+            }
 
             return false
         }
@@ -85,14 +102,14 @@ async function main(options: OptionValues, directory: string): Promise<boolean> 
 (async () => {
     program
         .description('Create Airsoft Database from csv files')
-        .option('-t, --test', 'test input files')
-        .option('-o, --output <output>', 'output file name', `${__dirname}/${REALM_FILE_NAME}`)
-        .argument('[directory]', 'directory containing csv files', `${__dirname}/../source_data/`)
+        .option('-t, --test', 'test only, no output')
+        .option('-o, --output <output>', 'output file name', path.normalize(`${__dirname}/../temp_realm_data/${REALM_FILE_NAME}`))
+        .argument('[directory]', 'directory containing csv files', path.normalize(`${__dirname}/../source_data/`))
         .action(async (directory) => {
             const options = program.opts();
             const test = options.test;
             const success = await main(options, directory);
-            console.log(`${success ? "Succeeded" : "FAILED"}`);
+            console.log(`${success ? "\nSucceeded" : "\nFAILED"}`);
             process.exit(success ? 0 : 1);
         });
